+++ initializer/main.go
@@ -10,7 +10,6 @@ import (
 	"os"
 	"sort"
 	"strings"
-	"time"
 )
 
 
@@ -19,10 +18,6 @@ const (
 	geckoImageNameEnvVar       = "GECKO_IMAGE_NAME"
 	chitSpammerImageNameEnvVar = "CHIT_SPAMMER_IMAGE_NAME"
 	defaultParallelism         = 4
-
-	// The max additional time we'll give to a test, on top of the per-test declared timeout, for setup & teardown
-	// TODO once we have an isBootstrapped endpoint that works, drop this down
-	additionalTestTimeoutBuffer = 300 * time.Second
 
 	// The number of bits to make each test network, which dictates the max number of services a test can spin up
 	// Here we choose 8 bits = 256 max services per test
@@ -141,7 +136,6 @@ func main() {
 			geckoImageNameEnvVar: *geckoImageNameArg,
 			chitSpammerImageNameEnvVar: *chitSpammerImageNameArg,
 		},
-		additionalTestTimeoutBuffer,
 		networkWidthBits)
 
 	// Create the container based on the configurations, but don't start it yet.